src: gen-index.py, raytracer.py
gen-index: fix a problem with leaf-nodes indexing referring to its
parent of the parent, instead of its parent.

raytracer: an experimental raytracer reading the binary format and
generating a small image to validate the concept (very slow). Changes to
the format may be validated using this raytracer before writing it in
WGSL.