index using axis-aligned box containers
Nodes with axis-aligned box containers can be easily raytraced.
Unlike gen-index, we allow modification of the mesh to achieve this.
Each node split the space by plane aligned with the axes,
splitting the triangles that intersect the splitting plane.

While the result will be more triangles, the node spaces will not be
intersecting, resulting in faster raytracing by limiting the maximum and
minimum depth of the ray.